Drop SWIG connection wrapper
